KFLUXINFRA-1719: Separate controller and webhook
For the controller we would need leader election.

We've seen the during etcd defrag, controllers which relies on leases
can restart.

This cause a problem when the webhook server runs in the same process
with the controller, since the webhook must be always available otherwise
pipelinerun creation will fail.

In addition, we would probably want to run different number of replicas
for the webhook (3) and the controller (2).

We are still going to build a single binary and a single container
image. The binary will have two sub commands for the controller and
webhook. This approach will optimize the build process.

The deployment manifest were adjusted and cleaned.
A dedicate service account with less privileges is used for the
webhook deployment.

e2e tested were adjusted as well to test the deployment of the
webhook server. Tests which checks the actual functionality weren't
changed in order to ensure this change doesn't introduce regressions.

Signed-off-by: Gal Ben Haim <gbenhaim@redhat.com>